feat(animation): PR1-4 — SampleTime invariants, TemporalSampleKey, temporal/spatial curve split, CubicBezier3D extensions, cinematic motion visual tests
PR1 (sample-time-invariants):
- Refactor SampleTime: single canonical coordinate (frame + FrameRate)
- Remove redundant seconds/fps fields; add derived seconds()/fps()/integral_frame()/fraction()
- Fix FrameContext: seconds()/progress() use effective_frame() instead of integer frame
- Update all callers: SceneBuilder, LayerBuilder, AnimatedCamera2_5D, CameraRig, etc.
- Add FPS precision tests (29.97 NTSC, 23.976 cinema, 25 PAL)

PR2 (temporal-cache-key):
- New TemporalSampleKey: {Frame frame, u32 subframe_tick, EvaluationVersion version}
- Replace SampleTimeKey with TemporalSampleKey in FrameCacheKey, NodeCacheKey, RenderFrameInfo
- Update frame_cache.cpp/node_cache.cpp digests for new key fields
- Add make_temporal_key() factory with tick carry and negative frame handling
- Static node key shared across 8 motion-blur sub-samples

PR3 (temporal-spatial-curve-split):
- New types: CubicBezier3D, TemporalCurve1D, SpatialKeyframe3D, MotionSegment3D,
  MotionPath3D, MotionPathSampler3D, PathTimingMode enum
- CubicBezier3D: position(u), derivative(u), sample(u), from_handles()
- TemporalCurve1D: wraps EasingCurve or AnimationCurve, maps t->progress

PR4 (bezier-auto-smooth-continuity):
- from_tangents(): construct bezier from tangent directions + magnitudes
- make_auto_smooth(): Catmull-Rom -> Bezier conversion with tension parameter
- is_c0_continuous_with(), is_c1_continuous_with(), is_geometric_c1_with()
- tangent_at(): unit-length direction with fallback to P0->P1 and -Z default
- TemporalCurve1D copy constructor/assignment for deep-copy of AnimationCurve

Cinematic motion visual test suite:
- 6 diagnostic visual tests: SubframeComb, ContactSheet, CacheParity,
  BezierHandles3D, ArcLengthSpacing, TemporalSpatialCurveSeparation
- Scene factories + compositing helpers (blit_into, contact_sheet_3x3, quad_2x2)
- Metric analysis functions with coefficient_of_variation, pixel_diff, clustering
- New CMake target: chronon3d_cinematic_motion_visual_tests
- Golden image support via CHRONON3D_UPDATE_GOLDENS=1

Registry clear() methods for clean test state reset across 7 registries.
ExtensionRegistry clear_modules() resets non-singleton registries + recreates builtins.
GraphExecutor plan cache switched to shared_ptr<const ExecutionPlan> for O(1) copies.

38 files changed, ~2800 lines added

+++ src/scene/presets/motion_resolver.cpp
@@ -83,7 +83,7 @@ MotionState resolve_motion_state(const FrameContext& ctx, const MotionObject& ob
     AnimationContext actx{
         ctx.frame,
         obj.time_value.end - obj.time_value.start,
-        ctx.fps()
+        static_cast<f32>(ctx.fps())
     };
     for (const auto& anim : obj.get_animations()) {
         anim.apply(actx, st);
